Fix: rm circular dependency (#4697)
* Fix: rm circular dependency

* Rm GPT bot
