--- .github/workflows/code-review-gpt.yml
@@ -1,25 +1,0 @@
-name: AI code review
-
-permissions:
-  contents: read
-  pull-requests: write
-
-on:
-  pull_request:
-    types: [opened]
-
-jobs:
-  review:
-    # Don't run on draft PRs
-    if: ${{ github.event.pull_request.draft == false }}
-    runs-on: ubuntu-latest
-    steps:
-      - uses: anc95/ChatGPT-CodeReview@main
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          OPENAI_API_KEY: ${{ secrets.OPENAI_API_KEY }}
-          # Optional
-          MODEL: gpt-4o # https://platform.openai.com/docs/models
-          PROMPT: |
-            Find potential bugs in these code changes.
-            If you cannot find any bugs, say nothing, just write a funny haiku summarizing the code changes and post ONLY that.
