+++ src/utils/transaction-guards.ts
@@ -69,7 +69,7 @@ import { isMultiSendCalldata } from './transaction-calldata'
 import { decodeMultiSendData } from '@safe-global/protocol-kit/dist/src/utils'
 import { OperationType } from '@safe-global/safe-core-sdk-types'
 import { LATEST_SAFE_VERSION } from '@/config/constants'
-import { extractMigrationL2MasterCopyAddress } from './safe-migrations'
+import { extractMigrationL2MasterCopyAddress } from '@/features/multichain/utils/extract-migration-data'
 
 export const isTxQueued = (value: TransactionStatus): boolean => {
   return [TransactionStatus.AWAITING_CONFIRMATIONS, TransactionStatus.AWAITING_EXECUTION].includes(value)
